// frontend/src/shared/api/generated/model/placeOrderRequest.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * StockQuest API
 * 모의 투자 챌린지 학습 플랫폼 API
 * OpenAPI spec version: 1.0.0
 */
import type { PlaceOrderRequestOrderType } from "./placeOrderRequestOrderType";
import type { PlaceOrderRequestSide } from "./placeOrderRequestSide";

export interface PlaceOrderRequest {
  /** 상품 키 */
  instrumentKey: string;
  /** 지정가 (지정가 주문 시 필수) */
  limitPrice?: number;
  /** 주문 유형 */
  orderType: PlaceOrderRequestOrderType;
  /**
   * 주문 수량
   * @minimum 0
   */
  quantity: number;
  /** 매수/매도 구분 */
  side: PlaceOrderRequestSide;
}