// frontend/src/shared/api/generated/model/placeOrderRequestSide.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * StockQuest API
 * Stock Trading Simulation Platform API
 * OpenAPI spec version: 1.0.0
 */

/**
 * 주문 방향
 */
export type PlaceOrderRequestSide =
  (typeof PlaceOrderRequestSide)[keyof typeof PlaceOrderRequestSide];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PlaceOrderRequestSide = {
  BUY: "BUY",
  SELL: "SELL",
} as const;