// frontend/src/shared/api/generated/model/placeOrderRequestOrderType.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * StockQuest API
 * Stock Trading Simulation Platform API
 * OpenAPI spec version: 1.0.0
 */

/**
 * 주문 유형
 */
export type PlaceOrderRequestOrderType =
  (typeof PlaceOrderRequestOrderType)[keyof typeof PlaceOrderRequestOrderType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PlaceOrderRequestOrderType = {
  MARKET: "MARKET",
  LIMIT: "LIMIT",
} as const;