// frontend/src/shared/api/generated/model/orderResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * StockQuest API
 * 모의 투자 챌린지 학습 플랫폼 API
 * OpenAPI spec version: 1.0.0
 */
import type { OrderResponseOrderType } from "./orderResponseOrderType";
import type { OrderResponseSide } from "./orderResponseSide";
import type { OrderResponseStatus } from "./orderResponseStatus";

export interface OrderResponse {
  executedAt?: string;
  executedPrice?: number;
  id?: number;
  instrumentKey?: string;
  limitPrice?: number;
  orderedAt?: string;
  orderType?: OrderResponseOrderType;
  quantity?: number;
  side?: OrderResponseSide;
  slippageRate?: number;
  status?: OrderResponseStatus;
}