// frontend/src/shared/api/generated/model/orderResponseStatus.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * StockQuest API
 * 모의 투자 챌린지 학습 플랫폼 API
 * OpenAPI spec version: 1.0.0
 */

export type OrderResponseStatus =
  (typeof OrderResponseStatus)[keyof typeof OrderResponseStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrderResponseStatus = {
  PENDING: "PENDING",
  EXECUTED: "EXECUTED",
  CANCELLED: "CANCELLED",
} as const;