// frontend/src/shared/api/generated/model/errorResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * StockQuest API
 * Stock Trading Simulation Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface ErrorResponse {
  /** 오류 코드 */
  error?: string;
  /** 오류 메시지 */
  message?: string;
  /** 요청 경로 */
  path?: string;
  /** HTTP 상태 코드 */
  status?: number;
  /** 오류 발생 시간 */
  timestamp?: string;
}